Incorporate variability in transition movement
Transitions differ across time. The configuration now expects an
addition column that represents the standard deviation of that
movement. This is used when calculating the number of people that
should move at a given time step, instead of it being a simple ratio.

Population is also assumed/allowed to vary for a given tile. The
average is now taken to deal with that.